Back/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.0.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.6
MarkupSafe==2.1.5
numpy==2.1.1
openapi-codec==1.3.2
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
simplejson==3.19.2
sqlparse==0.5.1
threadpoolctl==3.5.0
uritemplate==4.1.1
urllib3==2.2.2
